Update usize used in paginate to u64 (PR#789)

+++ examples/actix3_example/api/src/lib.rs
@@ -14,7 +14,7 @@ use serde::{Deserialize, Serialize};
 use std::env;
 use tera::Tera;
 
-const DEFAULT_POSTS_PER_PAGE: usize = 5;
+const DEFAULT_POSTS_PER_PAGE: u64 = 5;
 
 #[derive(Debug, Clone)]
 struct AppState {
@@ -23,8 +23,8 @@ struct AppState {
 }
 #[derive(Debug, Deserialize)]
 pub struct Params {
-    page: Option<usize>,
-    posts_per_page: Option<usize>,
+    page: Option<u64>,
+    posts_per_page: Option<u64>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
